build: refactor components
