+++ package.json
@@ -16,9 +16,10 @@
     "bootstrap": "pnpm i",
     "clean:lib": "rimraf lib && rimraf es && rimraf dist",
     "build": "sh scripts/build.sh",
+    "build:new": "gulp -f build/gulpfile.ts build",
     "build:helper": "esno build/build-helper.ts",
     "build:comps": "rimraf dist/components && esno build/components.ts",
-    "build:style": "gulp --cwd ./build",
+    "build:style": "gulp -f build/style/gulpfile.ts",
     "build:prod": "sh scripts/publish.sh",
     "build:directives": "cd packages/directives && pnpm clean && pnpm build",
     "build:hooks": "cd packages/hooks && pnpm clean && pnpm build",
@@ -90,6 +91,7 @@
     "@commitlint/config-conventional": "^13.1.0",
     "@pnpm/find-workspace-packages": "^3.1.13",
     "@pnpm/logger": "^4.0.0",
+    "@pnpm/types": "^7.4.0",
     "@rollup/plugin-alias": "^3.1.5",
     "@rollup/plugin-commonjs": "^15.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
