+++ scripts/build.sh
@@ -7,7 +7,9 @@ pnpm clean:lib
 # build all packages in case of error
 
 # build components
-pnpm build:comps
+# pnpm build:comps
+pnpm build:new
+
 rsync -a dist/types/components/ dist/element-plus/es/components/
 rsync -a dist/types/components/ dist/element-plus/lib/components/
